Extract comp to a separate file
Extract the `StepTemplate` component to a separate file and rename to
`BridgeProcessStep` so we can reuse them on the unminting details page.
